+++ timm/models/__init__.py
@@ -70,6 +70,6 @@ from .layers import TestTimePoolHead, apply_test_time_pool
 from .layers import convert_splitbn_model, convert_sync_batchnorm
 from .layers import is_scriptable, is_exportable, set_scriptable, set_exportable, is_no_jit, set_no_jit
 from .layers import set_fast_norm
-from ._pretrained import PretrainedCfg, filter_pretrained_cfg, generate_default_cfgs, split_model_name_tag
+from .pretrained import PretrainedCfg, filter_pretrained_cfg, generate_default_cfgs, split_model_name_tag
 from .registry import register_model, model_entrypoint, list_models, list_pretrained, is_model, list_modules,\
     is_model_in_modules, is_model_pretrained, get_pretrained_cfg, get_pretrained_cfg_value
